+++ spark/src/test/scala/org/apache/spark/sql/delta/DeltaTestUtils.scala
@@ -24,6 +24,7 @@ import java.util.concurrent.{ConcurrentHashMap, TimeUnit}
 import scala.collection.JavaConverters._
 import scala.collection.concurrent
 import scala.reflect.ClassTag
+import scala.reflect.runtime.universe._
 import scala.util.matching.Regex
 
 import com.databricks.spark.util.{Log4jUsageLogger, UsageRecord}
@@ -414,6 +415,36 @@ object DeltaTestUtils extends DeltaTestUtilsBase {
       dataChange: Boolean = true,
       stats: String = "{\"numRecords\": 1}"): AddFile = {
     AddFile(encodedPath, partitionValues, size, modificationTime, dataChange, stats)
+  }
+
+
+  /**
+   * Discovers all DeltaOperations.Operation subclasses using reflection.
+   * Returns a Set of operation class names.
+   *
+   * This is useful for tests that need to ensure exhaustive coverage of all operations.
+   */
+  def getAllDeltaOperations: Set[String] = {
+    val mirror = runtimeMirror(getClass.getClassLoader)
+    val moduleSymbol =
+      mirror.staticModule("org.apache.spark.sql.delta.DeltaOperations")
+    val moduleMirror = mirror.reflectModule(moduleSymbol)
+    val instance = moduleMirror.instance
+
+    val instanceMirror = mirror.reflect(instance)
+    val symbol = instanceMirror.symbol
+    val traitOperation =
+      typeOf[org.apache.spark.sql.delta.DeltaOperations.Operation].typeSymbol
+
+    symbol.typeSignature.members.flatMap {
+      case cls: ClassSymbol
+        if cls.isCaseClass && cls.isPublic && cls.toType.baseClasses.contains(traitOperation) =>
+        Some(cls.name.toString)
+      case obj: ModuleSymbol
+        if obj.isPublic && obj.moduleClass.asType.toType.baseClasses.contains(traitOperation) =>
+        Some(obj.name.toString)
+      case _ => None
+    }.toSet
   }
 
   /**
